+++ lib/src/elements/part/measure/note/note.dart
@@ -1,30 +1,54 @@
 import 'package:music_xml/src/elements/part/measure/note/unpitched/unpitched.dart';
 import 'package:xml/xml.dart';
 
+import 'accidental.dart';
+import 'beam.dart';
 import 'chord.dart';
-import 'grace/grace.dart';
+import 'dot.dart';
+import 'grace.dart';
+import 'notations/notations.dart';
 import 'pitch/pitch.dart';
+import '../duration.dart' as music_xml;
+import 'rest.dart';
+import 'staff.dart';
+import 'stem.dart';
+import '../../../voice.dart';
 import '../../../../basic_attributes.dart';
 import '../../../../local.dart';
-import '../../../../lyric.dart';
+import 'lyric.dart';
 import '../../../../music_xml_parser_state.dart';
 import '../../../../note_duration.dart';
-import '../../../../tie.dart';
+import 'tie.dart';
 
 /// Internal representation of a MusicXML `<note>` element.
 /// https://www.w3.org/2021/06/musicxml40/musicxml-reference/elements/note/
 class Note extends XmlElement {
+  // TODO: support attributes: attack, color, default-x, default-y, dynamics,
+  //       end-dynamics, font-family, font-size, font-style, font-weight, id,
+  //       pizzicato, print-dot, print-leger, print-lyric, print-object,
+  //       print-spacing, relative-x, relative-y, release, time-only
+  // TODO: support children: <cue>, <instrument>, <footnote>, <level>,
+  //       <notehead>, <notehead-text>, <play>, <listen>
   final Grace? grace;
   final Chord? chord;
   final Pitch? pitch;
   final Unpitched? unpitched;
+  final Rest? rest;
+  final music_xml.Duration? duration;
+  final Voice? voice;
+  final List<Dot> dots;
+  final List<Beam> beams;
+  final Stem? stem;
+  final Staff? staff;
+  final Accidental? accidental;
+  final List<Notations> notations;
 
   final int midiChannel;
   final int midiProgram;
   final int velocity;
-  final int voice;
-  final bool isRest;
   final NoteDuration noteDuration;
+
+  bool get isRest => rest != null;
 
   /// Tied notes will have the same note id.
   int get noteId => _noteId;
@@ -46,13 +70,18 @@ class Note extends XmlElement {
     Chord? chord;
     Pitch? pitch;
     Unpitched? unpitched;
-    var voice = 1;
-    var isRest = false;
-    String? duration;
-    var dots = 0;
+    Voice? voice;
+    Rest? rest;
+    music_xml.Duration? duration;
+    final dots = <Dot>[];
     String? type;
     double? tupletRatio;
 
+    final List<Beam> beams = [];
+    Stem? stem;
+    Staff? staff;
+    Accidental? accidental;
+    final List<Notations> notationsList = [];
     final List<Lyric> lyrics = [];
     List<Tie> ties = [];
 
@@ -70,40 +99,52 @@ class Note extends XmlElement {
         case Local.unpitched:
           unpitched = Unpitched.parse(child);
           break;
-        case 'duration':
-          duration = child.innerText;
+        case Local.duration:
+          duration = music_xml.Duration.parse(child);
           break;
-        case 'rest':
-          isRest = true;
+        case Local.rest:
+          rest = Rest.parse(child);
           break;
-        case 'voice':
-          voice = int.parse(child.innerText);
+        case Local.voice:
+          voice = Voice.parse(child);
           break;
-        case 'dot':
-          dots++;
+        case Local.dot:
+          dots.add(Dot.parse(child));
           break;
-        case 'type':
+        case Local.type:
           type = child.innerText;
           break;
-        case 'time-modification':
-          // A time-modification element represents a tuplet_ratio
+        case Local.timeModification:
           tupletRatio = _parseTuplet(child);
           break;
-        case 'lyric':
+        case Local.beam:
+          beams.add(Beam.parse(child));
+          break;
+        case Local.stem:
+          stem = Stem.parse(child);
+          break;
+        case Local.staff:
+          staff = Staff.parse(child);
+          break;
+        case Local.accidental:
+          accidental = Accidental.parse(child);
+          break;
+        case Local.notations:
+          notationsList.add(Notations.parse(child));
+          break;
+        case Local.lyric:
           lyrics.add(Lyric.parse(child, state));
           break;
-        case 'tie':
-          ties.add(Tie.parse(child, state));
+        case Local.tie:
+          ties.add(Tie.parse(child));
           break;
-        default:
-        // TODO: support remaining <note> child elements
       }
     }
     final noteDuration = NoteDuration.parse(
       chord != null,
       grace != null,
-      duration,
-      dots,
+      duration?.positiveDivisions,
+      dots.length,
       type,
       tupletRatio,
       state,
@@ -119,11 +160,18 @@ class Note extends XmlElement {
       chord,
       pitch,
       unpitched,
+      rest,
+      duration,
+      dots,
+      beams,
+      stem,
+      staff,
+      accidental,
+      notationsList,
       state.midiChannel,
       state.midiProgram,
       state.velocity,
       voice,
-      isRest,
       noteDuration,
       pitchMap,
       lyrics.isNotEmpty ? lyrics : null,
@@ -136,11 +184,18 @@ class Note extends XmlElement {
     this.chord,
     this.pitch,
     this.unpitched,
+    this.rest,
+    this.duration,
+    this.dots,
+    this.beams,
+    this.stem,
+    this.staff,
+    this.accidental,
+    this.notations,
     this.midiChannel,
     this.midiProgram,
     this.velocity,
     this.voice,
-    this.isRest,
     this.noteDuration,
     this.pitchMap,
     this.lyrics,
@@ -152,6 +207,15 @@ class Note extends XmlElement {
             if (chord != null) chord,
             if (pitch != null) pitch,
             if (unpitched != null) unpitched,
+            if (rest != null) rest,
+            if (duration != null) duration,
+            if (voice != null) voice,
+            ...dots,
+            ...beams,
+            if (stem != null) stem,
+            if (staff != null) staff,
+            if (accidental != null) accidental,
+            ...notations,
           ],
         );
 
@@ -168,10 +232,12 @@ class Note extends XmlElement {
   }
 
   /// Returns true if this note is not tied to a previous note
-  bool get isNoteOn => ties.isEmpty || ties.first.type != StartStop.stop;
+  bool get isNoteOn =>
+      ties.isEmpty || ties.first.type.startStop != StartStop.stop;
 
   /// Returns true if this note is not tied to a following note
-  bool get isNoteOff => ties.isEmpty || ties.last.type != StartStop.start;
+  bool get isNoteOff =>
+      ties.isEmpty || ties.last.type.startStop != StartStop.start;
 
   /// Returns true if this note is tied to a previous note
   bool get continuesOtherNote => !isNoteOn;